Fix an issue that was preventing nova to boot normal vms
The havana change to the allocate_for_instance method was
removing the device_owner attribute from the dict passed
as param when a port is created. Neutron Icehouse introduced
a new notification system that notifies nova when a port
has been created. The notification fails if the device_owner
attribute is not passed.

Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>

Change-Id: I0ffc4dd57c7c604464481cc6e39943529b87278f